requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ase==3.23.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.83
bleach==6.0.0
blinker==1.4
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cmake==3.26.4
colorlog==6.9.0
comm==0.1.3
confection==0.1.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.7
Cython==0.29.36
datasets==3.0.1
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
devscripts===2.22.1ubuntu1
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.8
docker-pycreds==0.4.0
e3nn==0.5.1
easydict==1.13
einops==0.6.1
exceptiongroup==1.1.2
execnet==1.9.0
executing==1.2.0
expecttest==0.1.3
fastjsonschema==2.17.1
filelock==3.12.2
flash-attn==2.4.3.post1
fonttools==4.40.0
fsspec==2023.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.21.0
google-auth-oauthlib==0.4.6
grpcio==1.56.0
httplib2==0.20.2
huggingface-hub==0.23.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
hypothesis==5.35.1
idna==3.4
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.14.7
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librosa==0.9.2
lightning-utilities==0.11.8
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
ml-collections==0.1.1
mock==5.0.2
more-itertools==8.10.0
mpmath==1.3.0
multiprocess==0.70.16
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1
notebook==6.4.10
numpy==1.22.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.27.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
pluggy==1.2.0
polygraphy==0.47.1
pooch==1.7.0
preshed==3.0.8
prettytable==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.11
pyg-lib==0.4.0+pt23cu121
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.4.0
pytest-flakefinder==1.1.0
pytest-rerunfailures==12.0
pytest-shard==0.1.2
pytest-xdist==3.3.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-hostlist==1.23.0
pytorch-quantization==2.1.2
PyYAML==6.0
pyzmq==25.1.0
rdkit-pypi==2022.9.3
referencing==0.29.1
regex==2023.6.3
requests==2.32.3
requests-oauthlib==1.3.1
resampy==0.4.2
rpds-py==0.8.8
rsa==4.9
rustworkx==0.15.1
safetensors==0.4.5
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sphinx-glpi-theme==0.3
spyrmsd==0.8.0
srsly==2.4.6
stack-data==0.6.2
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tbb==2021.9.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
timeout-decorator==0.5.0
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torch-ema==0.3
torch-geometric==2.6.0
torch_cluster==1.6.3+pt23cu121
torch_scatter==2.1.2+pt23cu121
torch_sparse==0.6.18+pt23cu121
torch_spline_conv==1.2.2+pt23cu121
torchaudio==2.3.0+cu121
torchmetrics==1.4.3
torchvision==0.18.0+cu121
tornado==6.3.2
tqdm==4.66.5
traitlets==5.9.0
transformers==4.28.1
Tree==0.2.4
triton==2.3.0
typer==0.9.0
types-dataclasses==0.6.6
typing==3.7.4.3
typing_extensions==4.12.0
ubuntu-pro-client==8001
unicore==0.0.1
urllib3==1.26.16
wadllib==1.3.6
wandb==0.17.0
wasabi==1.1.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xdoctest==1.0.2
xxhash==3.5.0
